chore: get rid of deprecated addSpanProcessor (#1126)
