+++ packages/web/src/types/config.ts
@@ -24,7 +24,7 @@ import { SplunkPostDocLoadResourceInstrumentationConfig } from '../SplunkPostDoc
 import { SocketIoClientInstrumentationConfig } from '../SplunkSocketIoClientInstrumentation'
 import { WebVitalsInstrumentationConfig } from '../webvitals'
 import { XMLHttpRequestInstrumentationConfig } from '@opentelemetry/instrumentation-xml-http-request'
-import { ReadableSpan } from '@opentelemetry/sdk-trace-base'
+import { ReadableSpan, SpanProcessor } from '@opentelemetry/sdk-trace-base'
 
 export interface SplunkOtelWebOptionsInstrumentations {
 	connectivity?: boolean | InstrumentationConfig
@@ -171,6 +171,8 @@ export interface SplunkOtelWebConfig {
 	 * @deprecated Renamed to rumAccessToken
 	 */
 	rumAuth?: string
+
+	spanProcessors?: Array<SpanProcessor>
 
 	/**
 	 * Config options passed to web tracer
